Run with java 8

+++ src/main/java/me/ildarorama/modbuscollector/support/DeviceResponse.java
@@ -4,68 +4,68 @@ import java.time.LocalDateTime;
 
 public class DeviceResponse {
     private LocalDateTime timestamp;
-    private double a1;
-    private double a2;
-    private double a3;
-    private double a4;
-    private double a5;
-    private double a6;
-    private double a7;
+    private float a1;
+    private float a2;
+    private float a3;
+    private float a4;
+    private float a5;
+    private float a6;
+    private float a7;
     private int a8;
 
-    public double getA1() {
+    public float getA1() {
         return a1;
     }
 
-    public void setA1(double a1) {
+    public void setA1(float a1) {
         this.a1 = a1;
     }
 
-    public double getA2() {
+    public float getA2() {
         return a2;
     }
 
-    public void setA2(double a2) {
+    public void setA2(float a2) {
         this.a2 = a2;
     }
 
-    public double getA3() {
+    public float getA3() {
         return a3;
     }
 
-    public void setA3(double a3) {
+    public void setA3(float a3) {
         this.a3 = a3;
     }
 
-    public double getA4() {
+    public float getA4() {
         return a4;
     }
 
-    public void setA4(double a4) {
+    public void setA4(float a4) {
         this.a4 = a4;
     }
 
-    public double getA5() {
+    public float getA5() {
         return a5;
     }
 
-    public void setA5(double a5) {
+    public void setA5(float a5) {
         this.a5 = a5;
     }
 
-    public double getA6() {
+    public float getA6() {
         return a6;
     }
 
-    public void setA6(double a6) {
+    public void setA6(float a6) {
         this.a6 = a6;
     }
 
-    public double getA7() {
+    public float getA7() {
         return a7;
     }
 
-    public void setA7(double a7) {
+    public void setA7(float a7) {
         this.a7 = a7;
     }
